+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/Block/utils/TemplateString/Variable/utils/useVariable.ts
@@ -9,15 +9,14 @@ import {
 
 import { VariableType } from "../../../../../../schema/primitives/variable";
 import { useQueryClient } from "@tanstack/react-query";
-import { z } from "zod";
 
 type UseVariableSuccess = [PossibleValues, undefined];
-type QueryFailure = [undefined, "queryIdNotFound" | "couldNotStringify"];
+export type VariableFailure = [undefined, "queryIdNotFound"];
 
-type UseVariableFailure =
+export type UseVariableFailure =
   | GetValueFromJsonPathFailure
   | ValidateVariableFailure
-  | QueryFailure;
+  | VariableFailure;
 
 /**
  * useVariable takes a variable string like "query.company-list.data.0.name" and
@@ -35,58 +34,25 @@ export const useVariable = (
     return [undefined, validationError];
   }
 
-  const { id, pointer } = variableObject;
-  const cacheKey = [id];
-  const queryState = queryClient.getQueryState(cacheKey);
+  const { id, pointer, namespace } = variableObject;
 
-  if (queryState === undefined) {
-    return [undefined, "queryIdNotFound"];
+  switch (namespace) {
+    case "query": {
+      const cacheKey = [id];
+      const queryState = queryClient.getQueryState(cacheKey);
+
+      if (queryState === undefined) {
+        return [undefined, "queryIdNotFound"];
+      }
+
+      const cachedData = queryClient.getQueryData(cacheKey);
+      const [data, error] = getValueFromJsonPath(cachedData, pointer);
+
+      if (error) {
+        return [undefined, error];
+      }
+
+      return [data, undefined];
+    }
   }
-
-  const cachedData = queryClient.getQueryData(cacheKey);
-  const [data, error] = getValueFromJsonPath(cachedData, pointer);
-
-  if (error) {
-    return [undefined, error];
-  }
-
-  return [data, undefined];
-};
-
-export const JSXValueSchema = z.union([
-  z.string(),
-  z.number(),
-  z.boolean(),
-  z.null(),
-  z.undefined(),
-]);
-
-export type JSXValueType = z.infer<typeof JSXValueSchema>;
-
-type UseVariableJSXSuccess = [JSXValueType, undefined];
-
-type UseVariableJSXFailure =
-  | GetValueFromJsonPathFailure
-  | ValidateVariableFailure
-  | QueryFailure;
-
-/**
- * useVariableJSX does the same as useVariable, but it will add some validation on top to ensure
- * that the value returned is JSX compatible.
- */
-export const useVariableJSX = (
-  value: VariableType
-): UseVariableJSXSuccess | UseVariableJSXFailure => {
-  const [data, error] = useVariable(value);
-
-  if (error) {
-    return [undefined, error];
-  }
-
-  const dataParsed = JSXValueSchema.safeParse(data);
-  if (!dataParsed.success) {
-    return [undefined, "couldNotStringify"];
-  }
-
-  return [dataParsed.data, undefined];
 };
